requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
yt-dlp>=2023.12.30
httpx>=0.26.0
mcstatus==11.1.1